Rename to dateTime.js
