+++ app/tests/unit/util/canProceed.test.js
@@ -3,7 +3,7 @@
  */
 const { get, set } = require('../../../src/services/datastore');
 const { canProceed } = require('../../../src/util/canProceed');
-const { dateTime } = require('../../../src/util/time');
+const { dateTime } = require('../../../src/util/dateTime');
 
 jest.mock('../../../src/services/datastore',
     () => ({
@@ -11,7 +11,7 @@ jest.mock('../../../src/services/datastore',
         get: jest.fn(),
         set: jest.fn(),
     }));
-jest.mock('../../../src/util/time');
+jest.mock('../../../src/util/dateTime');
 
 const datastoreGet = get;
 const datastoreSet = set;
